+++ webview-ui/src/components/settings/providers/Anthropic.tsx
@@ -1,0 +1,84 @@
+import { useCallback, useState } from "react"
+import { Checkbox } from "vscrui"
+import { VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
+
+import { ApiConfiguration } from "@roo/shared/api"
+
+import { useAppTranslation } from "@src/i18n/TranslationContext"
+import { VSCodeButtonLink } from "@src/components/common/VSCodeButtonLink"
+
+import { inputEventTransform, noTransform } from "../transforms"
+
+type AnthropicProps = {
+	apiConfiguration: ApiConfiguration
+	setApiConfigurationField: (field: keyof ApiConfiguration, value: ApiConfiguration[keyof ApiConfiguration]) => void
+}
+
+export const Anthropic = ({ apiConfiguration, setApiConfigurationField }: AnthropicProps) => {
+	const { t } = useAppTranslation()
+
+	const [anthropicBaseUrlSelected, setAnthropicBaseUrlSelected] = useState(!!apiConfiguration?.anthropicBaseUrl)
+
+	const handleInputChange = useCallback(
+		<K extends keyof ApiConfiguration, E>(
+			field: K,
+			transform: (event: E) => ApiConfiguration[K] = inputEventTransform,
+		) =>
+			(event: E | Event) => {
+				setApiConfigurationField(field, transform(event as E))
+			},
+		[setApiConfigurationField],
+	)
+
+	return (
+		<>
+			<VSCodeTextField
+				value={apiConfiguration?.apiKey || ""}
+				type="password"
+				onInput={handleInputChange("apiKey")}
+				placeholder={t("settings:placeholders.apiKey")}
+				className="w-full">
+				<label className="block font-medium mb-1">{t("settings:providers.anthropicApiKey")}</label>
+			</VSCodeTextField>
+			<div className="text-sm text-vscode-descriptionForeground -mt-2">
+				{t("settings:providers.apiKeyStorageNotice")}
+			</div>
+			{!apiConfiguration?.apiKey && (
+				<VSCodeButtonLink href="https://console.anthropic.com/settings/keys" appearance="secondary">
+					{t("settings:providers.getAnthropicApiKey")}
+				</VSCodeButtonLink>
+			)}
+			<div>
+				<Checkbox
+					checked={anthropicBaseUrlSelected}
+					onChange={(checked: boolean) => {
+						setAnthropicBaseUrlSelected(checked)
+
+						if (!checked) {
+							setApiConfigurationField("anthropicBaseUrl", "")
+							setApiConfigurationField("anthropicUseAuthToken", false)
+						}
+					}}>
+					{t("settings:providers.useCustomBaseUrl")}
+				</Checkbox>
+				{anthropicBaseUrlSelected && (
+					<>
+						<VSCodeTextField
+							value={apiConfiguration?.anthropicBaseUrl || ""}
+							type="url"
+							onInput={handleInputChange("anthropicBaseUrl")}
+							placeholder="https://api.anthropic.com"
+							className="w-full mt-1"
+						/>
+						<Checkbox
+							checked={apiConfiguration?.anthropicUseAuthToken ?? false}
+							onChange={handleInputChange("anthropicUseAuthToken", noTransform)}
+							className="w-full mt-1">
+							{t("settings:providers.anthropicUseAuthToken")}
+						</Checkbox>
+					</>
+				)}
+			</div>
+		</>
+	)
+}
