+++ webview-ui/src/components/settings/providers/Vertex.tsx
@@ -1,0 +1,97 @@
+import { useCallback } from "react"
+import { VSCodeLink, VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
+
+import { ApiConfiguration } from "@roo/shared/api"
+
+import { useAppTranslation } from "@src/i18n/TranslationContext"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@src/components/ui"
+
+import { inputEventTransform } from "../transforms"
+import { VERTEX_REGIONS } from "../constants"
+
+type VertexProps = {
+	apiConfiguration: ApiConfiguration
+	setApiConfigurationField: (field: keyof ApiConfiguration, value: ApiConfiguration[keyof ApiConfiguration]) => void
+}
+
+export const Vertex = ({ apiConfiguration, setApiConfigurationField }: VertexProps) => {
+	const { t } = useAppTranslation()
+
+	const handleInputChange = useCallback(
+		<K extends keyof ApiConfiguration, E>(
+			field: K,
+			transform: (event: E) => ApiConfiguration[K] = inputEventTransform,
+		) =>
+			(event: E | Event) => {
+				setApiConfigurationField(field, transform(event as E))
+			},
+		[setApiConfigurationField],
+	)
+
+	return (
+		<>
+			<div className="text-sm text-vscode-descriptionForeground">
+				<div>{t("settings:providers.googleCloudSetup.title")}</div>
+				<div>
+					<VSCodeLink
+						href="https://cloud.google.com/vertex-ai/generative-ai/docs/partner-models/use-claude#before_you_begin"
+						className="text-sm">
+						{t("settings:providers.googleCloudSetup.step1")}
+					</VSCodeLink>
+				</div>
+				<div>
+					<VSCodeLink
+						href="https://cloud.google.com/docs/authentication/provide-credentials-adc#google-idp"
+						className="text-sm">
+						{t("settings:providers.googleCloudSetup.step2")}
+					</VSCodeLink>
+				</div>
+				<div>
+					<VSCodeLink
+						href="https://developers.google.com/workspace/guides/create-credentials?hl=en#service-account"
+						className="text-sm">
+						{t("settings:providers.googleCloudSetup.step3")}
+					</VSCodeLink>
+				</div>
+			</div>
+			<VSCodeTextField
+				value={apiConfiguration?.vertexJsonCredentials || ""}
+				onInput={handleInputChange("vertexJsonCredentials")}
+				placeholder={t("settings:placeholders.credentialsJson")}
+				className="w-full">
+				<label className="block font-medium mb-1">{t("settings:providers.googleCloudCredentials")}</label>
+			</VSCodeTextField>
+			<VSCodeTextField
+				value={apiConfiguration?.vertexKeyFile || ""}
+				onInput={handleInputChange("vertexKeyFile")}
+				placeholder={t("settings:placeholders.keyFilePath")}
+				className="w-full">
+				<label className="block font-medium mb-1">{t("settings:providers.googleCloudKeyFile")}</label>
+			</VSCodeTextField>
+			<VSCodeTextField
+				value={apiConfiguration?.vertexProjectId || ""}
+				onInput={handleInputChange("vertexProjectId")}
+				placeholder={t("settings:placeholders.projectId")}
+				className="w-full">
+				<label className="block font-medium mb-1">{t("settings:providers.googleCloudProjectId")}</label>
+			</VSCodeTextField>
+			<div>
+				<label className="block font-medium mb-1">{t("settings:providers.googleCloudRegion")}</label>
+				<Select
+					value={apiConfiguration?.vertexRegion || ""}
+					onValueChange={(value) => setApiConfigurationField("vertexRegion", value)}>
+					<SelectTrigger className="w-full">
+						<SelectValue placeholder={t("settings:common.select")} />
+					</SelectTrigger>
+					<SelectContent>
+						{VERTEX_REGIONS.map(({ value, label }) => (
+							<SelectItem key={value} value={value}>
+								{label}
+							</SelectItem>
+						))}
+					</SelectContent>
+				</Select>
+			</div>
+		</>
+	)
+}
